PS/2 fixes: global on/off switch, button fixes...
spurious movements still not understood, but with global off switch it
is possible to continue using the keyboard.h

+++ src/Keyboard.h
@@ -100,7 +100,7 @@ uint8_t     g_mouse_keys;
 int8_t      g_mouse_modifier;
 uint16_t    g_mouse_mode_sum[4];
 uint8_t     g_tp_counter;
-uint32_t    g_mouse_mode;
+uint8_t     g_mouse_mode;
 uint32_t    g_mouse_lmb;
 uint8_t     g_mouse_double;
